refactor: move OpenSpiel game generation to a standalone script and update documentation
